+++ getopt.net.tests/GetOptTests_RealWorld.cs
@@ -144,6 +144,22 @@ namespace getopt.net.tests {
         }
 
         [TestMethod]
+        public void TestGetNextOptShortWithRequiredArgs_SeparatedBySpace()
+        {
+            var opt = new GetOpt();
+            opt.ShortOpts = "G:V:";
+            opt.AppArgs = new[] { "-G", "debug", "-V", "test" };
+
+            var optChar = (char)opt.GetNextOpt(out var optArg);
+            Assert.AreEqual('G', optChar);
+            Assert.AreEqual("debug", optArg);
+
+            optChar = (char)opt.GetNextOpt(out optArg);
+            Assert.AreEqual('V', optChar);
+            Assert.AreEqual("test", optArg);
+        }
+
+        [TestMethod]
         public void TestGetNextOptShort_WithFallbackToLongOpt() {
             var opt = new GetOpt();
             opt.Options = new[] {
@@ -236,15 +252,16 @@ namespace getopt.net.tests {
         }
 
         [TestMethod]
-        [ExpectedException(typeof(ParseException))]
-        public void TestFilenameOnly_ExpectException() {
+        public void TestFilenameOnly_DoNotIgnoreInvalidOpts() {
             var opt = new GetOpt();
             opt.ShortOpts = string.Empty;
             opt.Options = Array.Empty<Option>();
             opt.AppArgs = new[] { "filename.txt" };
             opt.IgnoreInvalidOptions = false;
 
-            opt.GetNextOpt(out var _); // Something expressing the existence of filename.txt should happen here.
+            var optChar = (char)opt.GetNextOpt(out var optArg);
+            Assert.AreEqual(GetOpt.InvalidOptChar, optChar);
+            Assert.AreEqual("filename.txt", optArg);
         }
 
         [TestMethod]
@@ -268,8 +285,7 @@ namespace getopt.net.tests {
         }
 
         [TestMethod]
-        [ExpectedException(typeof(ParseException))]
-        public void TestOptionBeforeFilename_ExpectException() {
+        public void TestOptionBeforeFilename_DoNotIgnoreInvalidOpts() {
             var opt = new GetOpt();
             opt.ShortOpts = "t";
             opt.Options = Array.Empty<Option>();
@@ -279,7 +295,9 @@ namespace getopt.net.tests {
             var optChar = (char)opt.GetNextOpt(out var optArg);
             Assert.AreEqual('t', optChar);
             Assert.IsNull(optArg);
-            opt.GetNextOpt(out var _); // Something expressing the existence of filename.txt should happen here.
+            optChar = (char)opt.GetNextOpt(out optArg);
+            Assert.AreEqual(GetOpt.InvalidOptChar, optChar);
+            Assert.AreEqual("filename.txt", optArg);
         }
 
         [TestMethod]
@@ -300,18 +318,20 @@ namespace getopt.net.tests {
         }
 
         [TestMethod]
-        [ExpectedException(typeof(ParseException))]
-        public void TestFilenameBeforeOptionGnuParsing_ExpectException() {
+        public void TestFilenameBeforeOptionGnuInOrderParsing_DoNotIgnoreInvalidOpts() {
             var opt = new GetOpt();
-            opt.ShortOpts = "t";
+            opt.ShortOpts = "-t";
             opt.Options = Array.Empty<Option>();
             opt.AppArgs = new[] { "filename.txt", "-t" };
             opt.IgnoreInvalidOptions = false;
 
             var optChar = (char)opt.GetNextOpt(out var optArg);
+            Assert.AreEqual(GetOpt.NonOptChar, optChar);
+            Assert.AreEqual("filename.txt", optArg);
+
+            optChar = (char)opt.GetNextOpt(out optArg);
             Assert.AreEqual('t', optChar);
             Assert.IsNull(optArg);
-            opt.GetNextOpt(out var _); // Something expressing the existence of filename.txt should happen here.
         }
 
         [TestMethod]
@@ -351,8 +371,7 @@ namespace getopt.net.tests {
         }
 
         [TestMethod]
-        [ExpectedException(typeof(ParseException))]
-        public void TestFilenameBeforeOptionPosixParsing_ExpectException() {
+        public void TestFilenameBeforeOptionPosixParsing_DoNotIgnoreInvalidOpts() {
             var opt = new GetOpt();
             opt.ShortOpts = "t";
             opt.Options = Array.Empty<Option>();
@@ -360,7 +379,12 @@ namespace getopt.net.tests {
             opt.IgnoreInvalidOptions = false;
 
             var optChar = (char)opt.GetNextOpt(out var optArg);
+            Assert.AreEqual(GetOpt.InvalidOptChar, optChar);
+            Assert.AreEqual("filename.txt", optArg);
+
             optChar = (char)opt.GetNextOpt(out optArg);
+            Assert.AreEqual('t', optChar);
+            Assert.IsNull(optArg);
         }
 
         [TestMethod]
@@ -620,12 +644,12 @@ namespace getopt.net.tests {
 
         [TestMethod]
         public void TestConvertLongOptsToShortOptsWithFailure() {
-            var emptyLongOpts = new Option[] {};
+            var emptyLongOpts = Array.Empty<Option>();
             var nullLongOpts = default(Option[]);
 
             Assert.IsNotNull(Extensions.ToShortOptString(emptyLongOpts));
             Assert.IsTrue(string.IsNullOrEmpty(Extensions.ToShortOptString(emptyLongOpts)));
-            Assert.IsNotNull(Extensions.ToShortOptString(nullLongOpts));
+            Assert.IsNotNull(Extensions.ToShortOptString(nullLongOpts!));
             Assert.IsTrue(string.IsNullOrEmpty(Extensions.ToShortOptString(emptyLongOpts)));
         }